// frontend/src/utils/automation-api-sdk/types.ts
/**
 *
 * (c) Copyright Ascensio System SIA 2024
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 */

export type Options = {
  docEditor: DocEditor;
  asc: Asc;
};

export type DocEditor = {
  createConnector: () => Connector;
  grabFocus: () => void;
}

export type Asc = {
  scope: { data: Object };
}

export type Connector = {
  callCommand: (callbck: () => void) => void;
  executeMethod: (name: string, args: Array<Object> | null, callbck: ((data: Object) => void) | null) => void;
}

export type FormData = {
  key: string
  value: string | boolean
}

export type BaseField = {
  key: string,
  tip: string,
  tag: string
}

export interface TextField extends BaseField {
  placeholder: string
}